+++ src/index-x.js
@@ -3,13 +3,13 @@ export * from './matrix';
 export * from './texture';
 export * from './atlas';
 export * from './pin';
-export * from './render';
 export * from './mouse';
+export * from './tween';
+export * from './util/math';
+
 export * from './root';
 export * from './canvas';
 export * from './sprite';
 export * from './anim';
 export * from './str';
 export * from './layout';
-export * from './tween';
-export * from './util/math';