+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * removed Stage.Math, added Stage.math with different function names
 * renamed sequence to align
 * removed Matrix reverse (use inverse)
+* removed Mouse.subscribe, replaced by new Mouse().mount
 
 #### v0.8.2
 * `render.js` renamed to `loop.js`